+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.2.0] - 2025-11-24
+
+### Changed
+
+- Deprecate `ReferenceProvider` in plugin API in favor of `MessageProvider` (existing implementations continue to work but should migrate).
+- Introduce `MessageProvider` in plugin API to supply archive data based on the full `Message` instead of just references and wire it into configuration and domain model handling.
+
 ## [10.1.0] - 2025-11-14
 
 ### Changed